internal/jujuapi: audit methods in JIMM facade
Implement access to the audit log in the JIMM facade. The API has been
implemented based on the future relational world, as has the dbmodel
definition, however the implementation uses the current audit log. This
causes a slight disconnect in the implementation. This will be resolved
when the relational data stores are fully implemented.

The current RPC methods are restricted to controller-admin users only.
This is more restrictive than the current HTTP based audit log. This
will be resolved when the relational data store is wired up.